chore: move custom hook `useOutsideClick` into react-components package
