+++ packages/react-components/src/hook/index.js
@@ -1,0 +1,7 @@
+import useOutsideClick from './use-outside-click'
+
+export { useOutsideClick }
+
+export default {
+  useOutsideClick,
+}
